# Make debugger significantly faster