--- a/src/libutil/pos-idx.hh
+++ b/src/libutil/pos-idx.hh
@@ -1,4 +1,5 @@
 #pragma once
+///@file
 
 #include <cinttypes>
 #include <functional>