@roberth 2025-05-11
Released in
- 2.24.13
- 2.27.0
- 2.28.0

Backported but end-of-life
- 2.26